requirements.txt
# requirements.txt
# A股量化交易系统依赖

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 数据接口
tushare>=1.2.89
akshare>=1.11.0

# 数据库
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0

# 时序数据库 (可选)
# influxdb-client>=1.36.0

# 回测框架
backtrader>=1.9.78
vectorbt>=0.24.0

# 技术分析
ta-lib>=0.4.28
pandas-ta>=0.3.14b

# 机器学习 (策略优化)
scikit-learn>=1.3.0
# lightgbm>=4.0.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0

# 日志
loguru>=0.7.0

# 配置管理
pyyaml>=6.0.0
python-dotenv>=1.0.0

# 网络请求
requests>=2.31.0
aiohttp>=3.8.0

# 工具
schedule>=1.2.0
colorama>=0.4.6
tqdm>=4.65.0
